[update] create signup page components
