+++ src/app/signup/components/index.ts
@@ -1,0 +1,1 @@
+export { SignUpForm } from './SignUpForm';
